pyproject.toml
[project]
name = "evm-voting-pytest"
version = "0.1.0"
description = "Simple Demo for testing telos evm stuff with python"
authors = [{ name = "Guillermo Rodriguez", email = "guillermo@telos.net" }]
requires-python = ">=3.11,<4"
readme = "README.md"
license = "MIT"
dependencies = [
    "rlp==4.1.0",
    "requests<2.32.0",
    "web3>=7.11.0",
    "py-leap",
    "docker>=7.1.0",
]

[dependency-groups]
dev = [
    "pdbp>=1.7.0",
    "pytest>=8.3.5",
]

[tool.uv]

[tool.uv.sources]
py-leap = { git = "https://github.com/openrepublic/py-leap.git", rev = 'v0.1a42' }
# py-leap = { path = "../py-leap", editable = true }

[tool.hatch.build.targets.sdist]
include = ["src/tevmtest"]

[tool.hatch.build.targets.wheel]
include = ["src/tevmtest"]

[tool.hatch.build.targets.wheel.sources]
"src/tevmtest" = "tevmtest"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = [
  '--show-capture=no',
]
markers = [
    'bootstrap: test requires full chain bootstrap',
    'contracts: specify contract you need deployed',
    'randomize: randomize eosio key and ports',
    'extra_plugins: enable extra nodeos plugins'
]
